change folder name and folderize Home component
