+++ src/component/ListUsers/style.scss
@@ -1,0 +1,1 @@
+@import '../../fonts.scss';
