+++ src/containers/Home/ChatArea/ChatSection.js
@@ -1,5 +1,5 @@
 import React, {useContext} from 'react';
-import ThemeContext from '../../contexts/ThemeContext';
+import ThemeContext from '../../../contexts/ThemeContext';
 import './ChatSection.scss';
 
 export const ChatSection = (props) => {
