+++ src/containers/Home/ChatArea/ChatSection.scss
@@ -1,4 +1,4 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
 .chat-wrapper {
   display: flex;
