+++ src/containers/Home/ChatArea/GroupDetail.scss
@@ -1,4 +1,4 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
 .groupDetail-context {
   transition: all 0.4s ease-in-out;
