+++ src/containers/Home/ChatArea/GroupDetailHeader.js
@@ -1,8 +1,8 @@
 import React, {useContext, useState} from 'react';
 import './GroupDetailHeader.scss';
-import groupAvt from '../../assets/images/groupAvt.png';
-import ThemeContext from '../../contexts/ThemeContext';
-import AvatarUpload from '../../component/AvatarUpload';
+import groupAvt from '../../../assets/images/groupAvt.png';
+import ThemeContext from '../../../contexts/ThemeContext';
+import AvatarUpload from '../../../component/AvatarUpload';
 
 export const GroupDetailHeader = ({chatTarget}) => {
   const theme = useContext(ThemeContext);
