+++ src/containers/Home/ChatArea/GroupDetailHeader.scss
@@ -1,4 +1,4 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
 .groupDetail-header {
   display: flex;
