+++ src/containers/Home/ChatArea/MessagesArea.js
@@ -1,6 +1,6 @@
 import React, {useContext} from 'react';
 import './MessagesArea.scss';
-import ThemeContext from '../../contexts/ThemeContext';
+import ThemeContext from '../../../contexts/ThemeContext';
 import ChatSection from './ChatSection';
 import GroupDetail from './GroupDetail';
