+++ src/containers/Home/ChatArea/MessagesArea.scss
@@ -1,4 +1,4 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
 .messagesArea-groupDetail {
   display: flex;
