+++ src/containers/Home/Header/Header.scss
@@ -1,4 +1,4 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
 
 .header-wrapper {
