+++ src/containers/Home/Menu/FavoriteSection.scss
@@ -1,7 +1,6 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
-
-.group-wrapper {
+.favorite-wrapper {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
@@ -10,41 +9,28 @@
   padding-bottom: 16px;
 }
 
-.group-wrapper>p {
+.favorite-wrapper>p {
   padding-bottom: 8px;
   white-space: nowrap;
   font-size: 1.5rem;
   cursor: default;
 }
 
-.group-wrapper>p>.fa-users {
+.favorite-wrapper>p>i {
   padding-left: 12px;
   padding-right: 8px;
 }
 
-.group-wrapper>p>.fa-plus {
-  padding: 0 4px 0 4px;
-  margin-left: 4px;
-  margin-right: 4px;
-  font-size: 1.1rem;
-  font-weight: 100;
-  vertical-align: baseline;
-  transition: all 0.2s ease-in-out;
-  cursor: pointer;
-}
-
-.group-wrapper>ul {
+.favorite-wrapper>ul {
   font-size: 1.1rem;
   padding-left: 12px;
-  word-wrap: break-word;
-  word-break: break-word;
   width: 100%;
 }
 
-.group-wrapper>ul>li {
+.favorite-wrapper>ul>li {
   display: flex;
   justify-content: flex-start;
-  align-items: center;
+  align-items: flex-start;
   word-wrap: break-word;
   word-break: break-word;
   padding-top: 6px;
