+++ src/containers/Home/Menu/GroupDialog.js
@@ -1,13 +1,13 @@
 import React, {useContext, useState} from 'react';
 import {useForm} from 'react-hook-form';
 import './GroupDialog.scss';
-import ThemeContext from '../../contexts/ThemeContext';
-import InputField from '../../component/InputField';
-import DialogButton from '../../component/DialogButton';
-import Dialog from '../../component/Dialog';
-import {post} from '../../utils/ApiCaller';
-import LocalStorageUtils from '../../utils/LocalStorageUtils';
-import {LOCALSTORAGE_TOKEN_NAME} from '../../configurations';
+import ThemeContext from '../../../contexts/ThemeContext';
+import InputField from '../../../component/InputField';
+import DialogButton from '../../../component/DialogButton';
+import Dialog from '../../../component/Dialog';
+import {post} from '../../../utils/ApiCaller';
+import LocalStorageUtils from '../../../utils/LocalStorageUtils';
+import {LOCALSTORAGE_TOKEN_NAME} from '../../../configurations';
 export const GroupDialog = (props) => {
   const [isFocused, setIsFocused] = useState(false);
   const theme = useContext(ThemeContext);
