+++ src/containers/Home/Menu/GroupDialog.scss
@@ -1,4 +1,4 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
 .dialogGroupBox {
   top: 0;
@@ -52,4 +52,3 @@
 .inputGroupName>.dialogBtn {
   margin-bottom: 8px;
 }
-
