+++ src/containers/Home/Menu/GroupSection.scss
@@ -1,7 +1,7 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
 
-.messages-wrapper {
+.group-wrapper {
   display: flex;
   flex-direction: column;
   align-items: flex-start;
@@ -10,19 +10,30 @@
   padding-bottom: 16px;
 }
 
-.messages-wrapper>p {
+.group-wrapper>p {
   padding-bottom: 8px;
   white-space: nowrap;
   font-size: 1.5rem;
   cursor: default;
 }
 
-.messages-wrapper>p>i {
+.group-wrapper>p>.fa-users {
   padding-left: 12px;
   padding-right: 8px;
 }
 
-.messages-wrapper>ul {
+.group-wrapper>p>.fa-plus {
+  padding: 0 4px 0 4px;
+  margin-left: 4px;
+  margin-right: 4px;
+  font-size: 1.1rem;
+  font-weight: 100;
+  vertical-align: baseline;
+  transition: all 0.2s ease-in-out;
+  cursor: pointer;
+}
+
+.group-wrapper>ul {
   font-size: 1.1rem;
   padding-left: 12px;
   word-wrap: break-word;
@@ -30,7 +41,7 @@
   width: 100%;
 }
 
-.messages-wrapper>ul>li {
+.group-wrapper>ul>li {
   display: flex;
   justify-content: flex-start;
   align-items: center;
