+++ src/containers/Home/Menu/UserNavbar.scss
@@ -1,4 +1,4 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
 
 .userNavbar-wrapper {
