+++ src/containers/Home/Menu/UserOption.js
@@ -1,8 +1,8 @@
 import React, {useContext, useState} from 'react';
 import './UserOption.scss';
-import ThemeContext from '../../contexts/ThemeContext';
-import {LOCALSTORAGE_TOKEN_NAME} from '../../configurations';
-import LocalStorageUtils from '../../utils/LocalStorageUtils';
+import ThemeContext from '../../../contexts/ThemeContext';
+import {LOCALSTORAGE_TOKEN_NAME} from '../../../configurations';
+import LocalStorageUtils from '../../../utils/LocalStorageUtils';
 
 const Options = (props) => {
   const theme = useContext(ThemeContext);
