+++ src/containers/Home/Menu/UserOption.scss
@@ -1,4 +1,4 @@
-@import '../../fonts.scss';
+@import '../../../fonts.scss';
 
 
 .options-wrapper {
